move "method should return a PromiEvent" back into logout.spec, break out testmode log out into its own suite
