+++ packages/@magic-sdk/provider/test/spec/modules/user/logout.spec.ts
@@ -1,5 +1,6 @@
 import browserEnv from '@ikscodes/browser-env';
-import { createMagicSDK, createMagicSDKTestMode } from '../../../factories';
+import { isPromiEvent } from '../../../../src/util';
+import { createMagicSDK } from '../../../factories';
 
 beforeEach(() => {
   browserEnv.restore();
@@ -16,13 +17,7 @@ test('Generate JSON RPC request payload with method `magic_auth_logout`', async 
   expect(requestPayload.params).toEqual([]);
 });
 
-test('If `testMode` is enabled, testing-specific RPC method is used', async () => {
-  const magic = createMagicSDKTestMode();
-  magic.user.request = jest.fn();
-
-  magic.user.logout();
-
-  const requestPayload = magic.user.request.mock.calls[0][0];
-  expect(requestPayload.method).toBe('magic_auth_logout_testing_mode');
-  expect(requestPayload.params).toEqual([]);
+test('method should return a PromiEvent', () => {
+  const magic = createMagicSDK();
+  expect(isPromiEvent(magic.user.logout())).toBeTruthy();
 });
